Added feature get node with free slots for harvesting
To harvest you need to connect to a node with free slots.
with this feature a button is provided when you are connected to a node
without free slots,
When the button is pressed the function will randomly choose a node
from the nodes provided and check if there is a slot free when there is
no slot free in that node, it will randomly choose another one. until
it found one or tried for more than 200 times.